Disambiguate the top and sub repo expanders
The classes do not share the same expander API and mypy cannot handle
the same variable being either of the two types. A Union or the base
class does not work as the API's do not match.

This is a little ugly in terms of branching, as we have no static
guarantee that the two variables are in scope when they are accessed,
and this could lead to runtime crashes. Just as the previous behavior
could initialize and then call different expanders and crash just the
same.