Cache the imported and exported commits
The cache stores the parents, submodule bumps, root tree id and
.gitmodules blob id for each imported commit. This means that the git
repository does not need to be accessed as much when loading the whole
history.

The cache also stores what submodules commit were expanded in each mono
repo commit and a mapping from toprepo to mono repo commit.

A new concept of WithoutCommitterId is introduced to be able to map back
to existing commits when splitting the commits into the sub
repositories.