Fix integration test
The previous move of the integration tests into a subdirectory silently
disabled all tests since it was missing a main.rs to make it a valid
crate. Normally cargo(?) generates a main.rs for every file found in
`tests/*` but this does not apply to crates, which need to define their
own entrypoints.

Once the tests were enabled they started breaking unless they were
limited to a single thread. This was caused by the directory
change in `workspace.rs` which isn't threadsafe, and has been fixed.