NILS: Prepare to allow fetch of super's own ssh urls.
We currently run the fetch logic on the data structure describing what
submodule paths to filter in (git-toprepo config).
Not a data structure that describes the filtered toprepo and all of its
projects (itself plus submodules). While this is a subtle difference
the data refinement ought to be refactored to be more correct.

This also ties into the dump subcommand for git-modules which might as
well print the fetch URL of each, but does not have the data available
either in the git-toprepo object.