Config: refactor the code to have clearer namspacing of git config keys
This introduces a runtime function to create namespaced git config keys.
It would be much better if it was a compilation time, const context,
function but that does not seem possible today. We could also use a
new-type but most of the uses are in Cmd::args in tests rather than our
own function calls, so that is more annoying, though it would really
give us the structural benefits of types.