Only warn for refs/remotes/ (#148)
No just refs/tags/ should considered as unfixable, also refs/notes/,
refs/pull/ and other unknown refs should not be reported. It is only
refs/remotes/ that is filtered and can be fixed.

Before this fix, invalid refs like 'refs/namespaces/foo/not-fetch-head'
would panic when trying to construct a gix::FullName from
'not-fetch-head' which is not a valid full name of a ref.