Keep the commit cache maps in sync (#160)
When the filtering failed, the cache mapping from mono commit object to
and from mono commit id had got new entries but not the top_to_mono_map.
Then, serialization of the cache failed.

This mitigates that problem by populating all three maps at the same
time, when the object and the id are both available.