Warn for empty submodule commit trees
git-submodule creates an empty directory for each submodule. In case a
submodule is substituted for an empty tree with git-toprepo, that
directory will disappear. User tooling might depend on the existance of
that empty directory, e.g. for each submodule mentioned in .gitmodules.

This commit prints a warning to make the user aware of the problem, so
that a dummy file can be added to the submodule as a fix.