Fix expansion of inner submodules
Inner submodules that were not updated when the super module was bumped
should still be checked out and not be reverted to a submodule. This
commit fixes that.

Example of the error:
top-1 -> subX-1 -> subY-1 contains subX/subY/file.txt
top-2 -> subX-2 -> subY-1 contains subX/subY as submodule

The code replaces the tree at subX which also resets subX/subY into a
submodule. Therefore, is subX is replaced by a tree, all inner
submodules also need to be replaced by trees even if they were not
changed.