Make BumpCache efficient enough to store all data (#235)
* Make BumpCache efficient enough to store all data

Instead of having the map keys as `(GitPath, RcKey<MonoRepoCommit>)`,
which allocates a string, use a double map to reduce the number of
stored `GitPaths`, i.e. `Map<GitPath, Map<RcKey<MonoRepoCommit>, T>>`.

For 1 million commits with 100 submodules and 50 bytes per key, this
should reduce the needed memory from 5 GiB to 800 MiB, so an LruCache is
not needed any more.

This also simplifies the code for the unwrapped recursive calls where
previously it was unknown if the data was left next time looking into
the storage.

* cargo rm lru v0.16.1