Only warn about problems at the tips of the top repo (#246)
There is no point of warning about problems in commits in the submodules
that are not referenced, it only makes sense to warn about problems that
are referenced at the tip of a top repo branch.

This commit duplicates code and is generally not very clean. A further
improvement is to interleave the load-and-fetch with the expand phases
so that only needed submodules are loaded and fetched. Let's put the
effort of cleanup then, but include tests to make sure the functionality
is there now and won't break during cleanup.