Remove 'git-toprepo fetch submod-path' with dubious use case
Keep `git-toprepo fetch` similar to `git fetch`:
    git fetch [<options>] [<repository> [<refspec>...]]
    git fetch [<options>] <group>
    git fetch --multiple [<options>] [(<repository> | <group>)...]
    git fetch --all [<options>]

Use a remote name or URL, but don't allow for a submodule path because
that is actually not a remote but rather the destination. In the future,
a command line `git-toprepo path-info` could be added to return URL,
server name etc. about a specific path in the repository.