Deduplicate body in commit messages when just footer diffs
When two commit messages only differs in the footer, e.g. with different
Change-Ids for Gerrit, merge the footers an keep only one copy of the
subject and body.

Before, deduplication only happened if the subject, body and footers
were the same. With this change, deduplication only requires the subject
and body to be equal.

NOTE: This change updates the expansion logic and therefore the monorepo
commit ids, so the import-cache version is bumped to enforce the
recombine operation.

Example: The following message...

    Subject line

    Change-Id: I61e0a73ad64816d09da88e5a0ed82412ee617484
    Git-Toprepo-Ref: subpath1 7837ff0aee85550d10a46bcb9a866f5167094a3d

    Subject line

    Change-Id: I403ccce60e75cad8c9d278b434d9bc17c45cfe1c
    Git-Toprepo-Ref: subpath2 c01ceee25ef214904b9ae0255ca349b318f0b3df

    Subject line

    Change-Id: I08424ce74bad63907e35df77b448faf6ed41e895
    Git-Toprepo-Ref: subpath3 48a8c249472166f5230fa67b00154ab8c7304954

    Git-Toprepo-Ref: <top> 03a281572f6f0f125eafac5d7d330ecce8bb6569

... becomes ...

    Subject line

    Git-Toprepo-Ref: <top> 03a281572f6f0f125eafac5d7d330ecce8bb6569
    Change-Id: I61e0a73ad64816d09da88e5a0ed82412ee617484
    Git-Toprepo-Ref: subpath1 7837ff0aee85550d10a46bcb9a866f5167094a3d
    Change-Id: I403ccce60e75cad8c9d278b434d9bc17c45cfe1c
    Git-Toprepo-Ref: subpath2 c01ceee25ef214904b9ae0255ca349b318f0b3df
    Change-Id: I08424ce74bad63907e35df77b448faf6ed41e895
    Git-Toprepo-Ref: subpath3 48a8c249472166f5230fa67b00154ab8c7304954